Refactor API call for patients into separate file
